+++ tests/Integration/PayPalTest.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace ArvPayoneApi\Integration;
+
+use ArvPayoneApi\Request\PaymentTypes;
+use ArvPayoneApi\Response\GenericResponse;
+
+/**
+ * Class PayPalTest
+ */
+class PayPalTest extends IntegrationTestAbstract
+{
+    public static function setUpBeforeClass()
+    {
+        parent::setUpBeforeClass();
+        self::$paymentMethod = PaymentTypes::PAYONE_PAY_PAL;
+    }
+
+    /**
+     * @depends testPreAuthSuccessfullyPlaced
+     * @group online
+     */
+    public function testCapturing(GenericResponse $preAuth)
+    {
+        $this->markTestSkipped('Customer action required on PayPal website');
+    }
+
+    /**
+     * @depends testAuthSuccessfullyPlaced
+     * @group online
+     */
+    public function testDebitAfterAuth(GenericResponse $auth)
+    {
+        $this->markTestSkipped('Will fail with "Desired status change not possible for this payment process" as Payone callbacks will not be processed.');
+    }
+
+}